adds support for libcrails-mongodb

+++ crails/scaffolds/templates/build/controller.cpp.cpp
@@ -17,6 +17,8 @@ public:
     parent_class(Crails::cast<string>(vars, "parent_class",  "ApplicationController")), 
     model_class(Crails::cast<string>(vars, "model_class",  "")), 
     model_header(Crails::cast<string>(vars, "model_header",  "")), 
+    id_type(Crails::cast<string>(vars, "id_type",  "")), 
+    database_backend(Crails::cast<string>(vars, "database_backend",  "")), 
     resource_scaffold(Crails::cast<bool>(vars, "resource_scaffold",  false)), 
     crud_scaffold(Crails::cast<bool>(vars, "crud_scaffold",  false))
   {}
@@ -24,9 +26,13 @@ public:
   void render()
   {
 ecpp_stream << "#include \"" << ( path );
-  ecpp_stream << ".hpp\"\n#include <crails/params.hpp>";
+  ecpp_stream << ".hpp\"\n#include <crails/params.hpp>\n#include <crails/http_response.hpp>";
  if (model_header.length() > 0){
-  ecpp_stream << "\n#include <crails/odb/to_vector.hpp>\n#include <boost/lexical_cast.hpp>\n#include \"lib/odb/application-odb.hxx\"\n#include \"" << ( model_header );
+  ecpp_stream << "\n#include <crails/odb/to_vector.hpp>\n#include <boost/lexical_cast.hpp>";
+   if (database_backend == "odb"){
+  ecpp_stream << "\n#include \"lib/odb/application-odb.hxx\"";
+ };
+  ecpp_stream << "\n#include \"" << ( model_header );
   ecpp_stream << "\"";
  };
   ecpp_stream << "\n\nusing namespace std;\n" << ( classname + "::" + classname );
@@ -43,12 +49,24 @@ ecpp_stream << "#include \"" << ( path );
  if (model_class.length() > 0){
   ecpp_stream << "\nshared_ptr<" << ( model_class );
   ecpp_stream << "> " << ( classname );
-  ecpp_stream << "::find_model(Crails::Odb::id_type id)\n{\n  database.find_one(model, odb::query<" << ( model_class );
-  ecpp_stream << ">::id == id);\n  vars[\"model\"] = model.get();\n  return model;\n}\n\nvoid " << ( classname );
-  ecpp_stream << "::require_model(Crails::Odb::id_type id)\n{\n  if (!model && !find_model(id))\n    respond_with(Crails::HttpStatus::not_found);\n}\n\nvoid " << ( classname );
-  ecpp_stream << "::find_list()\n{\n  odb::result<" << ( model_class );
+  ecpp_stream << "::find_model(" << ( id_type );
+  ecpp_stream << " id)\n{";
+ if (database_backend.length() > 0){
+  ecpp_stream << "\n  database.find_one(model, id);\n  vars[\"model\"] = model.get();";
+ };
+  ecpp_stream << "  return model;\n}\n\nvoid " << ( classname );
+  ecpp_stream << "::require_model(" << ( id_type );
+  ecpp_stream << " id)\n{\n  if (!model && !find_model(id))\n    respond_with(Crails::HttpStatus::not_found);\n}\n\nvoid " << ( classname );
+  ecpp_stream << "::find_list()\n{";
+ if (database_backend == "odb"){
+  ecpp_stream << "\n  odb::result<" << ( model_class );
   ecpp_stream << "> results;\n\n  database.find(results);\n  model_list = Crails::Odb::to_vector<" << ( model_class );
-  ecpp_stream << ">(results);\n  vars[\"models\"] = &model_list;\n}";
+  ecpp_stream << ">(results);";
+ }else if (database_backend == "mongodb"){
+  ecpp_stream << "\n  Crails::MongoDB::Result<" << ( model_class );
+  ecpp_stream << "> results;\n\n  database.find(results);\n  model_list = results.to_vector();";
+ };
+  ecpp_stream << "  vars[\"models\"] = &model_list;\n}";
  };
   ecpp_stream << "\n";
  if (resource_scaffold || crud_scaffold){
@@ -71,7 +89,11 @@ ecpp_stream << "#include \"" << ( path );
  if (model_class.length() > 0){
   ecpp_stream << "\n  " << ( model_class );
   ecpp_stream << " model;\n  model.edit(params[" << ( model_class );
-  ecpp_stream << "::scope]);\n  database.save(model);\n  redirect_to(\"" << ( router_path );
+  ecpp_stream << "::scope]);";
+ if (database_backend.length() > 0){
+  ecpp_stream << "\n  database.save(model);";
+ };
+  ecpp_stream << "  redirect_to(\"" << ( router_path );
   ecpp_stream << "/\" + boost::lexical_cast<std::string>(model.get_id()));";
  };
   ecpp_stream << "\n}\n";
@@ -84,19 +106,25 @@ ecpp_stream << "#include \"" << ( path );
   ecpp_stream << "::update()\n{";
  if (model_class.length() > 0){
   ecpp_stream << "\n  model->edit(params[" << ( model_class );
-  ecpp_stream << "::scope]);\n  database.save(*model);\n  redirect_to(\"" << ( router_path );
+  ecpp_stream << "::scope]);";
+ if (database_backend.length() > 0){
+  ecpp_stream << "\n  database.save(*model);";
+ };
+  ecpp_stream << "\n  redirect_to(\"" << ( router_path );
   ecpp_stream << "/\" + boost::lexical_cast<std::string>(model->get_id()));";
  };
   ecpp_stream << "\n}\n\nvoid " << ( classname );
   ecpp_stream << "::destroy()\n{";
- if (model_class.length() > 0){
+ if (model_class.length() > 0 && database_backend.length() > 0){
   ecpp_stream << "\n  database.destroy(*model);";
  };
   ecpp_stream << "\n  redirect_to(\"" << ( path );
   ecpp_stream << "\");\n}";
  };
   ecpp_stream << "";
-    this->target.set_body(ecpp_stream.str());
+    std::string _out_buffer = ecpp_stream.str();
+    _out_buffer = this->apply_post_render_filters(_out_buffer);
+    this->target.set_body(_out_buffer);
   }
 private:
   std::stringstream ecpp_stream;
@@ -107,6 +135,8 @@ private:
   string parent_class;
   string model_class;
   string model_header;
+  string id_type;
+  string database_backend;
   bool resource_scaffold;
   bool crud_scaffold;
 };
